+++ apps/spark/src/components/App/PageContext.tsx
@@ -4,15 +4,20 @@ import FooterNavigation from "../Footer/FooterNavigation";
 import Footer from "../Footer/Footer";
 import { metaDataForResponsiveDevices } from "../../utils/Preview/MetaData";
 import { StyledGrid } from "../PageGrid/PageGrid";
+import { BreadcrumbContextProvider } from "../../context/BreadcrumbContext";
+import Breadcrumb from "../Breadcrumb/Breadcrumb";
 
 const PageContext: FC = ({ children }) => {
   return (
-    <StyledGrid className={"cm-grid "} {...metaDataForResponsiveDevices()}>
-      <Header />
-      {children}
-      <FooterNavigation />
-      <Footer />
-    </StyledGrid>
+    <BreadcrumbContextProvider>
+      <StyledGrid className={"cm-grid "} {...metaDataForResponsiveDevices()}>
+        <Header />
+        <Breadcrumb />
+        {children}
+        <FooterNavigation />
+        <Footer />
+      </StyledGrid>
+    </BreadcrumbContextProvider>
   );
 };
 export default PageContext;