+++ apps/spark/src/components/RichText/DynamicComponent.tsx
@@ -1,6 +1,6 @@
 import React, { createElement } from "react";
 import RichTextElement, { RichTextElementProps } from "./RichTextElement";
-import { Attribute, transformAttributes } from "./RichTextHelper";
+import { Attribute, slugify, TOC_ELEMENTS, transformAttributes } from "./RichTextHelper";
 
 export interface DynamicComponentProps {
   name: string;
@@ -17,7 +17,15 @@ const DynamicComponent: React.FC<DynamicComponentProps> = ({ name, children, att
         })}
     </>
   );
-  return createElement(name, transformAttributes(attributes), childrenToRender);
+
+  // add id attribute for table of contents
+  const props: any = { ...transformAttributes(attributes) };
+  if (TOC_ELEMENTS.indexOf(name) >= 0 && children) {
+    const text = children[0].data || "";
+    props.id = slugify(text);
+  }
+
+  return createElement(name, props, childrenToRender);
 };
 
 export default DynamicComponent;